fix: favorite button in ItemDetails page
When the user was on the ItemDetails page and they clicked the favorite icon, it was updating on the home page and favorites page as well as sending a request to the backend but it just wasn't updating on the ItemDetails page that the user was on. Resolved by adding a separate local state for favorited item in SingleItem component to show the user. Also had to add a line in toggleFavorite function of search store to refresh the home page cause the home page wasn't updating. Will need to make this faster in the future cause right now it's rerendering the entire page.